Added Notification DB

+++ Db/Models/EventList.go
@@ -8,6 +8,6 @@ type EventList struct {
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
 	Events      []Event `json:"events"`
-	UserID      uint    `json:"userId"`
+	UserID      int64   `json:"userId"`
 	PriorityID  int64   `json:"priorityID"`
 }
